requirements.txt
httpx==0.28.1
schedule==1.2.2
python-dotenv==1.1.0